feat : Cafe -> Main class로 psvm 메서드 옮긴 후 CafeApplication 구현
