+++ src/main/java/org/example/application/Main.java
@@ -1,0 +1,9 @@
+package org.example.application;
+
+public class Main {
+
+    public static void main(String[] args) {
+        CafeApplication cafeApplication = new CafeApplication();
+        cafeApplication.run();
+    }
+}
